+++ PHP_PROJECTS/Blogging_site/views/admin/categoryDetails.php
@@ -5,7 +5,7 @@ if(isset($_SESSION['user_id'])){
 
     require_once('../../database/connection.php');
     
-    require_once('userDetailsVerify.php');
+    require_once('../common/userDetailsVerify.php');
 
     $userData = userVerification($_SESSION['user_id'] , $conn);
 
@@ -71,7 +71,7 @@ else{
                     <th>Delete</th>
                 </tr>
                 <?php 
-                    $sql = "SELECT *  FROM category ORDER BY created_at DESC";
+                    $sql = "SELECT *  FROM storycategory ORDER BY created_at DESC";
                     $result = mysqli_query($conn , $sql);
                     $userDataArray = mysqli_fetch_all($result , MYSQLI_ASSOC);
                     
@@ -85,7 +85,7 @@ else{
                                     <td>{$values['deleted_at']}</td>
                                     <td><a class='btn btn-primary' href='Editcategorydetails.php?category_id={$values['id']}' >Update</a></td>";
                                     if(empty($values['deleted_at'])){
-                                        echo "<td><a class='btn btn-danger' href='deletecategory.php?category_id={$values['id']}' >Delete</a></td>";   
+                                        echo "<td><a class='btn btn-danger' onclick=\"return confirm('Do you want to delete the category')\" href='deletecategory.php?category_id={$values['id']}' >Delete</a></td>";   
                                     }
                                 echo "</tr>";
                     }
